+++ src/sosi_origo_ne_ragel.cpp
@@ -2,7 +2,7 @@
 /* #line 1 "ragel/sosi_origo_ne.rl" */
 /*
  *  This file is part of the command-line tool sosicon.
- *  Copyright (C) 2012  Espen Andersen
+ *  Copyright (C) 2014  Espen Andersen, Norwegian Broadcast Corporation (NRK)
  *
  *  This is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -140,7 +140,7 @@ _resume:
 			else if ( (*p) > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (unsigned int)(_mid - _keys);
+				_trans += (_mid - _keys);
 				goto _match;
 			}
 		}
@@ -163,7 +163,7 @@ _resume:
 			else if ( (*p) > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += (unsigned int)((_mid - _keys)>>1);
+				_trans += ((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
